refactor(cart): extract CartShareCoordinator + CartChoiceDialog
Pull the Share picker + text/CSV/PDF handlers into their own class.
The 'title + explainer per row' dialog moves alongside as a top-level
extension so the clear-cart flow can use it too — dropping the private
duplicate off CartActivity.
